Due to the unit-testing issue, unfortunately I wasn't able to make much progress today. Oh well. I did get an in-browser unit test page up for the JavaScript VDF system, and got the system working (as far as it has been in Firefox) in Chrome also. This is helpful because of its better debugger tools. (apart from Firebug, anyway, which has issues with TypeScript, apparently)
Really looking forward to being done with the serializer porting. It was, and still is to some extent, fun; but it's been delayed a few days now due to inefficiencies/outside influences, and I'm starting to slow down. I just need one more solid push to ensure the project's completion, I think.